`optimize_lit_cmp`
