+++ crates/swc_ecma_minifier/src/compress/optimize/mod.rs
@@ -1836,19 +1836,6 @@ impl VisitMut for Optimizer<'_> {
             debug_assert_valid(e);
         }
 
-        if let Expr::Bin(bin) = e {
-            let expr = self.optimize_lit_cmp(bin);
-            if let Some(expr) = expr {
-                report_change!("Optimizing: Literal comparison");
-                self.changed = true;
-                *e = expr;
-            }
-        }
-
-        if e.is_seq() {
-            debug_assert_valid(e);
-        }
-
         self.compress_cond_expr_if_similar(e);
 
         if e.is_seq() {
